feat: clean up for embedded
