+++ src/app.module.ts
@@ -6,10 +6,10 @@ import { MongooseModule } from '@nestjs/mongoose';
 
 import { Definition, DryerModule } from '../lib';
 import { AuthResolver } from './resolvers';
-import { Product, Tag, User, Author, UserHook, Image, Customer, Variant } from './models';
+import { Product, Tag, User, Author, UserHook, Image, Customer, Variant, Computer } from './models';
 import { Ctx } from './ctx';
 
-const definitions: Definition[] = [Product, Tag, User, Author, Image, Variant, Customer];
+const definitions: Definition[] = [Product, Tag, User, Author, Image, Variant, Customer, Computer];
 
 @Module({
   imports: [
