+++ src/models/index.ts
@@ -3,3 +3,4 @@ export * from './user';
 export * from './product';
 export * from './variant';
 export * from './customer';
+export * from './computer';
